ReadMe.MD
This readme had to be created because this or a license was created for this project. I am releasing sooner than expected. Again, WARNING, it's EXPERIMENTAL CODE. You can use Windows 10's Network Reset Utility to attempt to recover. I've had to a couple of times but again, I am on a cellular connection. It's not stable like other types of connections.